refactor: extract room creation logic into custom hook
- Move room creation logic from CreateRoomPage into useCreate custom hook
- Export User type from auth-context for better type reusability
- Improve code organization and maintainability
- Reduce component complexity by separating business logic
